more styling, csv tools hooked up

+++ js/views/bottom_table.js
@@ -5,6 +5,9 @@ define([
   'jquery',     // lib/jquery/jquery
   'underscore', // lib/underscore/underscore
   'backbone',
+  'models/query',
+  'collections/queries',
+  'collections/tree_ids',
   'jquery_migrate',
   'jquery_drag',
   'jquery_core',
@@ -17,15 +20,8 @@ define([
   'slick_dataview',
   'slick_checkbox',
   'slick_selection',
-  'models/query',
-  'collections/queries',
-  'dataTables',
-  'tablesorter',
-  'metadata',
-  'tablecloth',
-  'lazyjson',
   'bootstrap'
- ], function($, _, Backbone, QueryModel, QueriesCollection, dataTables){
+ ], function($, _, Backbone, QueryModel, QueriesCollection, TreeIDCollection){
 	var BottomTableView = Backbone.View.extend({
 		el: '#data_table_container',
     model: QueryModel,
@@ -93,29 +89,19 @@ define([
     },
 
     clearSlickGrid: function(){
+      console.log("deleting all, removing all rectangles");
+      $("#clear_table").trigger("click"); // map view listens to this to remove rectangles
       this.data = [];//clear data
-
-      // $("#grid").empty();
-
-      // dataView = new Slick.Data.DataView();
-      // grid = new Slick.Grid("#grid", dataView, this.columns, this.options);
-
-
-      // this.grid.setSelectionModel(new Slick.RowSelectionModel());
-      // this.grid.registerPlugin(this.checkboxSelector);
-      //console.log(this.data);
       this.dataView.beginUpdate();
       this.dataView.setItems(this.data);
-      console.log(this.checkboxSelector);
       this.dataView.endUpdate();
 
       this.grid.updateRowCount();
       this.grid.render();
-
+      $(".slick-column-name input[type=checkbox]").attr('checked',false); //should do this automatically but it doesn't for some reason...
       this.dataView.syncGridSelection(this.grid, true);
-      // this.grid.resizeCanvas(); 
-      //reset selected count
-      $("#sample_count").html(0);
+      $("#sample_count").html(0); //reset selected count
+      this.sub_collection.reset(); //reset collection
     },
 
     updateSlickGrid: function(rectangleQuery){
@@ -124,12 +110,7 @@ define([
       var sts_is_query = encodeURIComponent(this.getTableQueryRectangle("sts_is",rectangleQuery));
       var tgdr_query = encodeURIComponent(this.getTableQueryRectangle("tgdr",rectangleQuery));
       var try_db_query = encodeURIComponent(this.getTableQueryRectangle("try_db",rectangleQuery));
-      console.log(sts_is_query);
-      console.log(tgdr_query);
-      console.log(try_db_query);
-
      
-      // $("#grid").empty();
       this.setLoaderIcon();
       $.ajax({
         url : 'getFusionMarkers.php',
@@ -144,13 +125,6 @@ define([
           that.data = that.data.concat($.map(response,function(a,i){
             return that.toObj(a,i);
           }));
-
-
-          // dataView = new Slick.Data.DataView();
-          // grid = new Slick.Grid('#grid', dataView, that.columns, that.options);
-
-          // grid.setSelectionModel(new Slick.RowSelectionModel());
-          // grid.registerPlugin(that.checkboxSelector);
 
           var sortCol = undefined;
           var sortDir = true;
@@ -180,18 +154,20 @@ define([
 
           that.dataView.syncGridSelection(that.grid, true);
 
-      	  that.grid.onSelectedRowsChanged.subscribe(function(){ 
+      	  that.grid.onSelectedRowsChanged.subscribe(function(){  // update selected count and set the sub collection to the selected ids
       		  $("#sample_count").html(that.grid.getSelectedRows().length);
+            that.sub_collection.reset();//remove all previous ids
+            $.each(that.grid.getSelectedRows(), function(index,idx){ //add newly selected ones
+              var id = that.dataView.getItemByIdx(idx)["id"].replace(/\.\d+$/,"");
+              var lat = that.dataView.getItemByIdx(idx)["lat"]; //lat and lng for world_clim tool
+              var lng = that.dataView.getItemByIdx(idx)["lng"];
+              that.sub_collection.add({
+                id: id,
+                lat: lat,
+                lng: lng
+              }); 
+            });
       	  });
-
-          // $("#remove_selected").on("click", function(){//use backbone events
-          //   var ids = that.dataView.mapRowsToIds(that.grid.getSelectedRows());
-          //   $.each(ids,function(index,id){
-          //     console.log("deleting: "+id);
-          //     that.dataView.deleteItem(id);
-          //   });
-          //   that.grid.invalidate();
-          // });
 
           console.log("Total samples: "+that.grid.getDataLength());
         }
@@ -222,27 +198,33 @@ define([
 		}
 	},
 
-	initialize: function(){
+	initialize: function(options){
+    this.sub_collection = options.sub_collection; //allows passing of tree_ids between this view and the tabs view
+    var that = this;
     this.initColumns();
     this.initGrid();
 		this.clearSlickGrid();//clear at first
 		this.collection.on('add change remove',this.refreshTable,this); //possibly reset?
-	},
+  },
 
   removeSelected: function(){
     var that = this;
-    var ids = this.dataView.mapRowsToIds(this.grid.getSelectedRows());
-    $.each(ids,function(index,id){
-      console.log("deleting: "+id);
-      that.dataView.deleteItem(id);
-    });
-    this.grid.invalidate();
+    if (this.grid.getSelectedRows().length == this.grid.getDataLength()){ // clear if all selected, not iterate remove
+      this.clearSlickGrid();
+    }
+    else{
+      var ids = this.dataView.mapRowsToIds(this.grid.getSelectedRows());
+      $.each(ids,function(index,id){
+        console.log("deleting: "+id);
+        that.dataView.deleteItem(id);
+      });
+      this.grid.invalidate();
+    }
   },
 
             
   events:{
     "click #remove_selected": "removeSelected",
-    "click #clear_table": "clearTable",
   }
 
 	});
